refactor: user schams & remove  fixme code
