+++ src/backend/app/users/user_deps.py
@@ -1,17 +1,9 @@
-from app.users.user_crud import get_user_by_email
+from app.users.user_crud import verify_token
 from fastapi import HTTPException, Request, Header
 from app.config import settings
 from app.users.auth import Auth
 from app.users.user_schemas import AuthUser
 from loguru import logger as log
-import jwt
-from typing import Any
-from passlib.context import CryptContext
-from psycopg import Connection
-from app.db import db_models
-from pydantic import EmailStr
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 
 async def init_google_auth():
@@ -58,42 +50,3 @@ async def login_required(
         raise HTTPException(status_code=401, detail="Access token not valid") from e
 
     return AuthUser(**user)
-
-
-def verify_token(token: str) -> dict[str, Any]:
-    """Verifies the access token and returns the payload if valid.
-
-    Args:
-        token (str): The access token to be verified.
-
-    Returns:
-        dict: The payload of the access token if verification is successful.
-
-    Raises:
-        HTTPException: If the token has expired or credentials could not be validated.
-    """
-    try:
-        return jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
-    except jwt.ExpiredSignatureError as e:
-        raise HTTPException(status_code=401, detail="Token has expired") from e
-    except Exception as e:
-        raise HTTPException(status_code=401, detail="Could not validate token") from e
-
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
-
-
-def get_password_hash(password: str) -> str:
-    return pwd_context.hash(password)
-
-
-async def authenticate(
-    db: Connection, email: EmailStr, password: str
-) -> db_models.DbUser | None:
-    db_user = await get_user_by_email(db, email)
-    if not db_user:
-        return None
-    if not verify_password(password, db_user["password"]):
-        return None
-    return db_user
