+++ src/backend/app/users/user_routes.py
@@ -142,7 +142,9 @@ async def my_data(
 ):
     """Read access token and get user details from Google"""
     user_info = await user_schemas.DbUser.get_or_create_user(db, user_data)
-    has_user_profile = await user_deps.get_userprofile_by_userid(db, user_info.id)
+    has_user_profile = await user_schemas.DbUserProfile.get_userprofile_by_userid(
+        db, user_info.id
+    )
     user_info_dict = user_info.model_dump()
     user_info_dict["has_user_profile"] = bool(has_user_profile)
     return user_info_dict
